Organized common code into the ta_quolab package
- Core functionality in ta_quolab.api
- Complete move query_catalog() to shared module
- Move constant data to new module
