+++ .bumpversion.cfg
@@ -7,8 +7,7 @@ commit = True
 search = version = {current_version}
 replace = version = {new_version}
 
-# XXX: Move this to a common layer name (or, even better, remove bin.d layers compleetly)
-[bumpversion:file:bin.d/50-modinput-timeline/quolab_ta.py]
+[bumpversion:file:bin/ta_quolab/__init__.py]
 search = __version__ = "{current_version}"
 replace = __version__ = "{new_version}"
