+++ bin/quolab_query.py
@@ -1,171 +1,24 @@
 #!/usr/bin/env python
 # coding=utf-8
 
-from __future__ import absolute_import, division, print_function, unicode_literals
+from __future__ import (absolute_import, division, print_function,
+                        unicode_literals)
 
-import os
-import sys
-import re
 import json
-import time
+import os
+import re
+import sys
 
 sys.path.insert(0, os.path.join(os.path.dirname(__file__), "..", "lib"))  # nopep8
 
-import six
-import requests
 from cypresspoint.datatype import as_bool
 from cypresspoint.searchcommand import ensure_fields
-from cypresspoint.spath import splunk_dot_notation
-from requests.auth import AuthBase, HTTPBasicAuth
-from requests.utils import default_user_agent
 from splunklib.client import Entity, HTTPError
-from splunklib.searchcommands import dispatch, GeneratingCommand, Configuration, Option, validators
-
-from quolab_ta import monotonic, __version__
-
-""" http debug logging
-import logging
-from http.client import HTTPConnection  # py3
-
-log = logging.getLogger('urllib3')
-log.setLevel(logging.DEBUG)
-"""
-
-quolab_classes = {
-    'fact': [
-        'autonomous-system',
-        'certificate',
-        'domain',
-        'email',
-        'envelope',
-        'file',
-        'function',
-        'hostname',
-        'import-table',
-        'ip-address',
-        'mutex',
-        'process',
-        'registry-key',
-        'url',
-        'wallet',
-        'export-table',
-        'malware',
-        'blob',
-        'ttp',
-        'organization',
-        'persona',
-        'region',
-        'tor-descriptor',
-        'transaction',
-        'yara-rule',
-    ],
-    'reference': [
-        'accesses',
-        'contains',
-        'creates',
-        'identified-as',
-        'loads',
-        'matches',
-        'relates-to',
-        'signed-by',
-        'receives-from',
-        'sends-to',
-        'delivered',
-        'resolved-to',
-    ],
-    'annotation': [
-        'attribute',
-        'text',
-        'interpreted-as',
-        'known-as',
-        'geodata',
-        'report',
-    ],
-    'sysfact': [
-        'case',
-        'resource',
-        'script',
-        'tag',
-        'timeline',
-        'text',
-        'user',
-        'group',
-        'subscription',
-        'connector',
-        'regulator',
-        'endpoint',
-    ],
-    'sysref': [
-        'queued',
-        'scheduled',
-        'executed',
-        'canceled',
-        'failed',
-        'observed-by',
-        'commented-by',
-        'monitors',
-        'associated-with',
-        'encases',
-        'tagged',
-        'uses',
-        'synchronized-with',
-        'implies',
-        'authorizes',
-        'member-of',
-        'produced',
-    ]
-}
-
-
-facets = [
-    'cases',
-    'contributors',
-    'document.magic',
-    'commented',
-    'document',
-    'sources',
-    'producers',
-    'refcount',
-    'vault-stored',
-    'display',
-    'actions',
-    'endpoints',
-    'latest-internal-observation',
-    'tagged',
-]
-
-quolab_types = set()
-quolab_class_from_type = {}
-
-resolve_override = {
-    "text": "sysfact",
-}
-
-
-def init():
-    for class_, types in quolab_classes.items():
-        for type_ in types:
-            if type_ in quolab_types:
-                if type_ in resolve_override:
-                    class_ = resolve_override[type_]
-                else:
-                    # XXX: This assertion would be better as a unittest
-                    raise AssertionError("Duplicate entry for {}:  {} vs {}".format(
-                        type_, quolab_class_from_type[type_], class_))
-            quolab_types.add(type_)
-            quolab_class_from_type[type_] = class_
-
-
-init()
-
-
-class QuolabAuth(AuthBase):
-    def __init__(self, token):
-        self._token = token
-
-    def __call__(self, request):
-        request.headers['Authorization'] = "Quoken {}".format(self._token)
-        return request
+from splunklib.searchcommands import (Configuration, GeneratingCommand, Option,
+                                      dispatch, validators)
+from ta_quolab import __version__
+from ta_quolab.api import QuoLabAPI
+from ta_quolab.const import facets, quolab_class_from_type, quolab_types
 
 
 @Configuration()
@@ -228,7 +81,7 @@ class QuoLabQueryCommand(GeneratingCommand):
     run_in_preview = False
 
     def __init__(self):
-        self.session = requests.Session()
+        self.quolab_api = None
         self.api_url = None
         self.api_username = None
         self.verify = True
@@ -291,6 +144,13 @@ class QuoLabQueryCommand(GeneratingCommand):
                             "from {} without secret.".format(self.api_url),
                             "Missing secret.  Did you run setup?")
 
+        # Setup quolab interface
+        self.quolab_api = QuoLabAPI(self.api_url, verify=self.verify)
+        if self.api_username == "<TOKEN>":
+            self.quolab_api.login_token(self.api_secret)
+        else:
+            self.quolab_api.login(self.api_username, self.api_secret)
+
     @classmethod
     def _order_to_dict(cls, order_option, query):
         """ Build the order portion of a query document given a dot-notation
@@ -321,101 +181,14 @@ class QuoLabQueryCommand(GeneratingCommand):
         """ Handle the query to QuoLab API that drives this SPL command
         Returns [results]
         """
-        def monotonic():
-            # PY3:  Switch to time.monotonic()
-            return time.time()
-
-        # CATALOG QUERY
-        session = self.session
-
-        start = monotonic()
-        # Allow total run time to be 10x the individual query limit
-        expire = start + (self.api_timeout * 10)
-
-        url = "{}/v1/catalog/query".format(self.api_url)
-        headers = {
-            'content-type': "application/json",
-            'user-agent': "ta-quolab/{} {}".format(__version__, default_user_agent())
-        }
-        # XXX: Per query limit optimization?  Maybe based on type, or number of facets enabled?
-        query["limit"] = min(query_limit, self.api_fetch_count)
-
-        # Q: What do query results look like when time has been exceeded?  Any special handling required?
-        query.setdefault("hints", {})["timeout"] = self.api_timeout
-        i = http_calls = 0
-
-        if self.api_username == "<TOKEN>":
-            auth = QuolabAuth(self.api_secret)
-        else:
-            auth = HTTPBasicAuth(self.api_username, self.api_secret)
-
-        while True:
-            data = json.dumps(query)
-            self.logger.debug("Sending query to API:  %r   headers=%r auth=%s",
-                              data, headers, auth.__class__.__name__)
-            try:
-                response = session.request(
-                    "POST", url,
-                    data=data,
-                    headers=headers,
-                    auth=auth,
-                    verify=self.verify)
-                http_calls += 1
-            except requests.ConnectionError as e:
-                self.logger.error("QuoLab API failed due to %s", e)
-                self.write_error("QuoLab server connection failed to {}", url)
-                return
-
-            if response.status_code >= 400 and response.status_code < 500:
-                body = response.json()
-                if "status" in body or "message" in body:
-                    status = body.get("status", response.status_code)
-                    message = body.get("message", "")
-                    self.logger.error("QuoLab API returned unexpected status response from query.  "
-                                      "status=%r message=%r query=%r", status, message, query)
-                    self.write_error("QuoLab query failed:  {} ({})", message, status)
-                    return
-
-            # If a non-success exit code was returned, and the resulting object doesn't have message/status, then just raise an exception.
-            try:
-                response.raise_for_status()
-                body = response.json()
-            except Exception as e:
-                self.logger.debug("Body response for %s:   %s", e, response.text)
-                raise
-
-            self.logger.debug("Response body:   %s", body)
-
-            records = body["records"]
-            for record in body["records"]:
-                result = (splunk_dot_notation(record))
-                result["_raw"] = json.dumps(record)
-                # Q:  Are there ever fields that should be returned as _time instead of system clock time?
-                result["_time"] = time.time()
-                yield result
-                i += 1
-                if i >= query_limit:
-                    break
-
-            if monotonic() > expire:
-                self.logger.warning("Aborting query due to time expiration")
-                break
-
-            if i >= query_limit:
-                break
-
-            ellipsis = body.get("ellipsis", None)
-            if ellipsis:
-                self.logger.debug("Query next batch.  i=%d, query_limit=%d, limit=%d, ellipsis=%s",
-                                  i, query_limit, query["limit"], ellipsis)
-                query["resume"] = ellipsis
-            else:
-                break
-        self.logger.info("Query/return efficiency: http_calls=%d, query_limit=%d, per_post_limit=%d duration=%0.3f",
-                         http_calls, query_limit, query["limit"], monotonic()-start)
+        # Return generator function
+        return self.quolab_api.query_catalog(query, query_limit,
+                                             timeout=self.api_timeout,
+                                             fetch_count=self.api_fetch_count,
+                                             write_error=self.write_error)
 
     def generate(self):
-        # Because the splunklib search interface does a *really* bad job a reporting exceptions / logging stack traces :-(
+        # Because the splunklib search interface does a bad job a reporting exceptions / logging stack traces :-(
         try:
             return self._generate()
         except Exception as e:
@@ -426,7 +199,6 @@ class QuoLabQueryCommand(GeneratingCommand):
             sys.exit(1)
 
     def _generate(self):
-        session = requests.Session()
 
         if self.mode == "advanced":
             query = self.query.replace("'", '"')
@@ -437,8 +209,7 @@ class QuoLabQueryCommand(GeneratingCommand):
                 self.write_error("Invalid JSON:  {}", e)
                 sys.exit(1)
 
-            # Handle scenario where top-level 'query' is provided.  Otherwise, assume given content should be placed under 'query'
-            # The QuoLab API handles this either way, but to add limit and facets an explicit 'query' must be present.
+            # Force use of a top-level 'query' for consistant querying
             if "query" not in query:
                 query = {"query": query}
 
@@ -477,82 +248,6 @@ class QuoLabQueryCommand(GeneratingCommand):
         results = self._query_catalog(query, self.limit)
 
         return ensure_fields(results)
-
-
-def build_searchbnf(stream=sys.stdout):
-    """
-    >>> import quolab_query
-    >>> quolab_query.build_searchbnf()
-    """
-    stream.write("[quolab-types]\n")
-    types = []
-    # Don't bother showing reference/sysref in the UI docs (to keep the list
-    # from becoming too long and unreadable)
-    for class_ in ["sysfact", "fact", "annotation"]:
-        types.extend(quolab_classes[class_])
-    stream.write("syntax = ({})\n".format("|".join(types)))
-
-    stream.write("[quolab-facets]\n")
-    stream.write("syntax = ({})\n".format("|".join(sorted(facets))))
-
-
-def build_facets(data):
-    """ Data from facets-serves.json
-
-    Data structure:
-    { "services" : [
-        {"(type)": "quolab...ClassName",
-        "id": "<name>"}
-     ] }
-    """
-    deprecated = {"casetree", "cache-id", "indirect"}
-    services = [service["id"] for service in data["services"]
-                if service["id"] not in deprecated]
-    return services
-
-
-def build_types(data):
-    """ Extract datetypes from model-types.json
-    API:    /v1/catalog/model/types
-
-    Data Structure:
-    { "types: {
-        "<class>: [
-            {"type": "<type>"}
-        ]
-    }}
-
-    data = json.load(open("model-types.json"))
-    build_types(data)
-    """
-    deprecated = {"ipynb", "misp-blob", "source", "sink", "task", "sighted-by"}
-    qlc = {}
-    for class_, types in data["types"].items():
-        type_names = [t["type"] for t in types if t["type"] not in deprecated]
-        qlc[class_] = type_names
-    # print(repr(qlc))
-    return qlc
-
-
-def build_from_json(output=sys.stdout):
-    # TODO:  Make list output show up on the following line.   Something like:  replace(": ['", ": [\n    '").  Maybe just use json dump?
-    from pprint import pprint
-    pp_args = {
-        "compact": False
-    }
-    if sys.version_info > (3, 8):
-        pp_args["sort_dicts"] = False
-    data = json.load(open("model-types.json"))
-    qlc = build_types(data)
-    output.write("quolab_classes = ")
-    pprint(qlc, stream=output, **pp_args)
-    output.write("\n\n")
-
-    data = json.load(open("facet-services.json"))
-    facets = build_facets(data)
-    output.write("facets = ")
-    pprint(facets, stream=output, **pp_args)
-    output.write("\n\n")
 
 
 if __name__ == '__main__':
