+++ bin/ta_quolab/__init__.py
@@ -1,0 +1,6 @@
+""" QuoLab Add on for Splunk
+"""
+
+# Assumption:  All sys.path fixes have already been applied by the importer
+
+__version__ = "0.10.0"
